working on the landing page backend logic

+++ admin/delete-category.php
@@ -6,12 +6,17 @@ if (isset($_GET['id'])) {
 
   // FOR LATER
   // update category_id of of posts that belong to this category to id uncategorized category
+  $update_query = "UPDATE posts SET category_id=6 WHERE category_id = $id";
+  $update_result = mysqli_query($connection, $update_query);
 
-  // delete category
-  $query = "DELETE FROM categories WHERE id = $id LIMIT 1";
-  $result = mysqli_query($connection, $query);
+  if (!mysqli_errno($connection)) {
+    // delete category
+    $query = "DELETE FROM categories WHERE id = $id LIMIT 1";
+    $result = mysqli_query($connection, $query);
 
-  $_SESSION['delete-category-success'] = "category deleted successfully";
+    $_SESSION['delete-category-success'] = "category deleted successfully";
+  }
+  
 }
 header('location: ' . ROOT_URL . 'admin/manage-categories.php');
 die();
